Added fast game mode without defined time intervals
Added fast game mode but without the defined time intervals ( which can be added in line 95). The interval found there is for pure testing only. The pieces created during this game mode are deleted when you loose and you may continue playing depending on your number of lives. Each time a live is discounted, the score gained during this game-mode will be subtracted.